requirements.txt
# E-Commerce Sales Performance & Customer Behavior Analysis
# Python Dependencies
# Last Updated: December 18, 2024

# Core Data Processing
pandas==2.0.3
numpy==1.24.3

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Jupyter Environment
jupyter==1.0.0
notebook==7.0.2
ipykernel==6.25.1

# Data Validation (Optional but recommended)
great-expectations==0.17.23
pandera==0.16.1

# Configuration Management
pyyaml==6.0.1

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Code Quality
black==23.9.1
flake8==6.1.0
isort==5.12.0

# Performance
tqdm==4.66.1

# Utilities
python-dateutil==2.8.2
openpyxl==3.1.2  # For Excel exports
